notifications list loaded by graphql
