+++ packages/app-extension/src/components/common/Layout/Router.tsx
@@ -33,7 +33,6 @@ import { AnimatePresence } from "framer-motion";
 
 import { Apps } from "../../Unlocked/Apps";
 import { Balances } from "../../Unlocked/Balances";
-import { Notifications } from "../../Unlocked/Balances/Notifications";
 import { RecentActivity } from "../../Unlocked/Balances/RecentActivity";
 import { Token } from "../../Unlocked/Balances/TokensWidget/Token";
 import { ChatDrawer } from "../../Unlocked/Messages/ChatDrawer";
@@ -42,6 +41,7 @@ import { Nfts } from "../../Unlocked/Nfts";
 import { NftsCollection } from "../../Unlocked/Nfts/Collection";
 import { NftChat, NftsExperience } from "../../Unlocked/Nfts/Experience";
 import { NftOptionsButton, NftsDetail } from "../../Unlocked/Nfts/NftDetail";
+import { Notifications } from "../../Unlocked/Notifications";
 import { SettingsButton } from "../../Unlocked/Settings";
 
 import { NavBackButton, WithNav } from "./Nav";
